login and register pages are linked properly
