+++ src/app/Http/Controllers/Auth/LoginController.php
@@ -14,7 +14,7 @@ class LoginController extends Controller
 
     public function index()
     {
-        return view('auth.login');
+        return view('loginpage');
     }
 
     public function store(Request $request)
