+++ src/app/Http/Controllers/Auth/RegisterController.php
@@ -18,7 +18,7 @@ class RegisterController extends Controller
 
     public function index()
     {
-        return view('auth.register');
+        return view('registerpage');
     }
 
     public function store(Request $request)
