+++ src/resources/views/auth/login.blade.php
@@ -1,6 +1,7 @@
 
 @section('login')
         <div class="container" style="display: flex; flex-direction: column; align-items: center;">
+            <p>Please Log in Here</p>
             @if (session('status'))
                 <div class="bg-red-500 p-4 rounded-lg mb-6 text-white text-center">
                     {{ session('status') }}
@@ -10,27 +11,36 @@
             <form action="{{ route('login') }}" method="post">
                 @csrf
 
-
-                <div class="mb-3">
-                    <label for="email" class="form-label">Email</label>
-                    <input type="text" name="email" id="email" placeholder="Your email" class="form-control" style="width: 200px;">
-
-                </div>
-
-                <div class="mb-3">
-                    <label for="password" class="sr-only">Password</label>
-                    <input type="password" name="password" id="password" placeholder="Your password" class="form-control" style="width: 200px;">
+                <div >
+                    <input type="text" name="email" id="email" placeholder="Your email" class="form-control form-width @error('email') border border-danger @enderror" value="{{old('email')}}">
+                    @error('email')
+                    <div class="text-danger mt-2 text-sm">
+                        {{ $message }}
+                    </div>
+                @enderror
 
                 </div>
 
                 <div class="mb-3">
-                    <div class="flex items-center">
+                    <input type="password" name="password" id="password" placeholder="Your password" class="form-control form-width @error('password') border border-danger @enderror">
+                    @error('password')
+                    <div class="text-danger mt-2 text-sm">
+                        {{ $message }}
+                    </div>
+                    @enderror
+                </div>
+
+                <div class="mb-3">
                         <input type="checkbox" name="remember" id="remember" class="mr-2">
                         <label for="remember">Remember me</label>
-                        <button type="submit" class="btn btn-primary mb-3">Submit</button>
-                    </div>
+                </div>
+                <div >
+                    <button type="submit" class="btn btn-primary btn-block mb-3" style="width: 300px!important;">Log in</button>
                 </div>
 
+                <div style="display: flex; flex-direction: column; align-items: center;">
+                    <p class>Don't have an account? <a href="{{route('register')}}">Sign Up</a> </p>
+                </div>
             </form>
     </div>
 @endsection('login')
