+++ src/resources/views/auth/register.blade.php
@@ -1,65 +1,58 @@
 
     <div class="flex justify-center">
         <div class="w-4/12 bg-white p-6 rounded-lg">
+
+            <p>Please Fill the information below to register</p>
+
             <form action="{{ route('register') }}" method="post">
                 @csrf
-                <div class="mb-4">
-                    <label for="name" class="sr-only">Name</label>
-                    <input type="text" name="name" id="name" placeholder="Your name" class="bg-gray-100 border-2 w-full p-4 rounded-lg @error('name') border-red-500 @enderror" value="{{ old('name') }}">
-
+                <div class>
+                    <input type="text" name="name" id="name" placeholder="Your Name" class="form-control form-width @error('name') border border-danger @enderror" value="{{ old('name') }}">
                     @error('name')
-                        <div class="text-red-500 mt-2 text-sm">
-                            {{ $message }}
-                        </div>
-                    @enderror
-                </div>
-
-                <div class="mb-4">
-                    <label for="username" class="sr-only">Username</label>
-                    <input type="text" name="username" id="username" placeholder="Username" class="bg-gray-100 border-2 w-full p-4 rounded-lg @error('username') border-red-500 @enderror" value="{{ old('username') }}">
-
-                    @error('username')
-                        <div class="text-red-500 mt-2 text-sm">
-                            {{ $message }}
-                        </div>
-                    @enderror
-                </div>
-
-                <div class="mb-4">
-                    <label for="email" class="sr-only">Email</label>
-                    <input type="text" name="email" id="email" placeholder="Your email" class="bg-gray-100 border-2 w-full p-4 rounded-lg @error('email') border-red-500 @enderror" value="{{ old('email') }}">
-
-                    @error('email')
-                        <div class="text-red-500 mt-2 text-sm">
-                            {{ $message }}
-                        </div>
-                    @enderror
-                </div>
-
-                <div class="mb-4">
-                    <label for="password" class="sr-only">Password</label>
-                    <input type="password" name="password" id="password" placeholder="Choose a password" class="bg-gray-100 border-2 w-full p-4 rounded-lg @error('password') border-red-500 @enderror" value="">
-
-                    @error('password')
-                        <div class="text-red-500 mt-2 text-sm">
-                            {{ $message }}
-                        </div>
-                    @enderror
-                </div>
-
-                <div class="mb-4">
-                    <label for="password_confirmation" class="sr-only">Password again</label>
-                    <input type="password" name="password_confirmation" id="password_confirmation" placeholder="Repeat your password" class="bg-gray-100 border-2 w-full p-4 rounded-lg @error('password_confirmation') border-red-500 @enderror" value="">
-
-                    @error('password_confirmation')
-                        <div class="text-red-500 mt-2 text-sm">
+                        <div class="text-danger mt-2 text-sm">
                             {{ $message }}
                         </div>
                     @enderror
                 </div>
 
                 <div>
-                    <button type="submit" class="bg-blue-500 text-white px-4 py-3 rounded font-medium w-full">Register</button>
+                    <input type="text" name="username" id="username" placeholder="Username" class="form-control form-width @error('username') border border-danger @enderror" value="{{ old('username') }}">
+                    @error('username')
+                        <div class="text-danger mt-2 text-sm">
+                            {{ $message }}
+                        </div>
+                    @enderror
+                </div>
+
+                <div>
+                    <input type="text" name="email" id="email" placeholder="Your email" class="form-control form-width @error('username') border border-danger @enderror" value="{{ old('email') }}">
+                    @error('email')
+                        <div class="text-danger mt-2 text-sm">
+                            {{ $message }}
+                        </div>
+                    @enderror
+                </div>
+
+                <div>
+                    <input type="password" name="password" id="password" placeholder="Choose a password" class="form-control form-width @error('password') border border-danger @enderror" value="">
+                    @error('password')
+                        <div class="text-danger mt-2 text-sm">
+                            {{ $message }}
+                        </div>
+                    @enderror
+                </div>
+
+                <div class = "mb-3">
+                    <input type="password" name="password_confirmation" id="password_confirmation" placeholder="Repeat your password" class="form-control form-width @error('password_confirmation') border border-danger @enderror" value="">
+                    @error('password_confirmation')
+                        <div class="text-danger mt-2 text-sm">
+                            {{ $message }}
+                        </div>
+                    @enderror
+                </div>
+
+                <div >
+                    <button type="submit" class="btn btn-primary btn-block mb-3" style="width: 300px!important;">Register</button>
                 </div>
             </form>
         </div>
